Complete code checking with cppling and valgrind for ex0

+++ exercises/exercise00/ex0.c
@@ -1,22 +1,21 @@
+// Copyright 2025 Tsung-Wei Chin
 // Name: Miles TsungWei Chin
 // Email: twchin@uw.edu
-
-// Recursive Pi calculator based on Nilakantha series
+// Subject: Pi calculator based on Nilakantha series
 
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 
-// Function declaration for recursive Nilakantha series 
+// Function declarations
 double nilakantha(int n);
 int is_input_valid(int argc, char const *argv[]);
 
-int main(int argc, char const *argv[])
-{
+int main(int argc, char const *argv[]) {
   if (!is_input_valid(argc, argv)) {
     return EXIT_FAILURE;
   }
-  
+
   int n = strtol(argv[1], NULL, 10);
   double pi_estimate = 0.0;
   for (int i = 0; i <= n; i++) {
@@ -40,7 +39,6 @@ double nilakantha(int n) {
   }
 }
 
-// Specific input check function
 int is_input_valid(int argc, char const *argv[]) {
   // Check whether args count is correct
   if (argc != 2) {
@@ -50,7 +48,7 @@ int is_input_valid(int argc, char const *argv[]) {
 
   // Check whether the input is a non-negative integer
   char *endptr;
-  long val = strtol(argv[1], &endptr, 10);
+  int val = strtol(argv[1], &endptr, 10);
   if (*endptr != '\0' || val < 0) {
     printf("Invalid input\n");
     return 0;
